refactor SqlGenerator to include unique terms
