+++ src/main/java/bio/terra/cda/app/generators/SubjectSqlGenerator.java
@@ -2,7 +2,7 @@ package bio.terra.cda.app.generators;
 
 import bio.terra.cda.generated.model.Query;
 
-@QueryGenerator(
+@EntityGeneratorData(
     entity = "subject",
     hasFiles = true,
     defaultOrderBy = "subject_id",
@@ -11,7 +11,7 @@ import bio.terra.cda.generated.model.Query;
     aggregatedFieldsSelectString = {
         "json_agg(distinct (subject_identifier.system, subject_identifier.field_name, subject_identifier.value)::system_data) as subject_identifier",
         "json_agg(distinct subject_associated_project.associated_project) AS subject_associated_project"})
-public class SubjectSqlGenerator extends SqlGenerator {
+public class SubjectSqlGenerator extends EntitySqlGenerator {
   public SubjectSqlGenerator(Query rootQuery, boolean filesQuery) {
     super(rootQuery, filesQuery);
   }
