+++ src/test/java/bio/terra/cda/app/controller/QueryApiControllerTest.java
@@ -53,7 +53,7 @@ class QueryApiControllerTest {
             });
 
     var expected =
-        "SELECT DISTINCT sex FROM subject LEFT JOIN subject_identifier AS subject_identifier ON subject.id = subject_identifier.subject_id WHERE system = 'GDC' ORDER BY sex";
+        "SELECT DISTINCT sex FROM subject LEFT JOIN subject_identifier AS subject_identifier ON subject.id = subject_identifier.subject_id WHERE system = 'GDC' ORDER BY sex  LIMIT 100";
     var result =
         mvc.perform(
                 post("/api/v1/unique-values")
